requirements.txt
PyYAML
aiohttp
aiosqlite
pytest
pytest-asyncio
